quic: move quic flags to use Envoy Runtime API.
