+++ source/common/quic/platform/http2_flags_impl.h
@@ -8,7 +8,8 @@
 
 #include "common/quic/platform/flags_impl.h"
 
-#define GetHttp2ReloadableFlagImpl(flag) quiche::FLAGS_quic_reloadable_flag_##flag->value()
+#define GetHttp2ReloadableFlagImpl(flag)                                                           \
+  Envoy::Runtime::runtimeFeatureEnabled(EnvoyReloadableFeature(flag))
 
 #define SetHttp2ReloadableFlagImpl(flag, value)                                                    \
   quiche::FLAGS_quic_reloadable_flag_##flag->setValue(value)
