+++ source/common/runtime/runtime_features.h
@@ -22,12 +22,8 @@ public:
 
   // This tracks config-guarded code paths, to determine if a given
   // runtime-guarded-code-path has the new code run by default or the old code.
-  bool enabledByDefault(absl::string_view feature) const {
-    return enabled_features_.find(feature) != enabled_features_.end();
-  }
-  bool existsButDisabled(absl::string_view feature) const {
-    return disabled_features_.find(feature) != disabled_features_.end();
-  }
+  bool enabledByDefault(absl::string_view feature) const;
+  bool existsButDisabled(absl::string_view feature) const;
 
 private:
   friend class RuntimeFeaturesPeer;
