+++ test/common/runtime/runtime_impl_test.cc
@@ -192,6 +192,12 @@ TEST_F(DiskLoaderImplTest, All) {
   // test_feature_false is not in runtime_features.cc and so is false by default.
   EXPECT_EQ(false, snapshot->runtimeFeatureEnabled("envoy.reloadable_features.test_feature_false"));
 
+  // Test that quic flags can be overriden.
+  EXPECT_EQ(true, snapshot->runtimeFeatureEnabled(
+                      "envoy.reloadable_features.quic_testonly_default_false"));
+  EXPECT_EQ(false, snapshot->runtimeFeatureEnabled(
+                       "envoy.reloadable_features.quic_testonly_default_true"));
+
   // Deprecation
   EXPECT_EQ(false, snapshot->deprecatedFeatureEnabled(
                        "envoy.deprecated_features.deprecated.proto:is_deprecated_fatal", false));
@@ -258,7 +264,7 @@ TEST_F(DiskLoaderImplTest, All) {
 
   EXPECT_EQ(0, store_.counter("runtime.load_error").value());
   EXPECT_EQ(1, store_.counter("runtime.load_success").value());
-  EXPECT_EQ(25, store_.gauge("runtime.num_keys", Stats::Gauge::ImportMode::NeverImport).value());
+  EXPECT_EQ(27, store_.gauge("runtime.num_keys", Stats::Gauge::ImportMode::NeverImport).value());
   EXPECT_EQ(4, store_.gauge("runtime.num_layers", Stats::Gauge::ImportMode::NeverImport).value());
 }
